requirements.txt
# Core Dependencies (Python 3.12 compatible)
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Google Cloud
google-cloud-documentai>=2.21.0
google-cloud-storage>=2.10.0

# Machine Learning (Python 3.12 compatible versions)
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0
joblib>=1.3.2

# Database (Python 3.12 compatible)
sqlalchemy>=2.0.25
alembic>=1.13.1

# Utilities (Python 3.12 compatible)
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dateutil>=2.8.2
pytesseract>=0.3.10
opencv-python>=4.8.1.78

# Development & Testing (Python 3.12 compatible)
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2

# Frontend (Python 3.12 compatible)
streamlit>=1.28.2
pillow>=10.1.0

# LangChain & AI (Python 3.12 compatible)
langchain>=0.1.0
langchain-google-vertexai>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.22
sentence-transformers>=2.2.2

# Automation & Monitoring Dependencies (Python 3.12 compatible)
watchdog>=3.0.0          # File system monitoring
aiofiles>=23.2.0         # Async file operations
schedule>=1.2.0          # Task scheduling
psutil>=5.9.6            # System monitoring
aiohttp>=3.9.1           # Async HTTP client
requests>=2.31.0        # HTTP requests for testing

# Additional Testing Dependencies (Python 3.12 compatible)
pytest-cov>=4.1.0       # Coverage testing
pytest-xdist>=3.5.0     # Parallel testing
pytest-html>=4.1.1      # HTML test reports
pytest-json-report>=1.5.0  # JSON test reports

# Security & Code Quality (Python 3.12 compatible)
bandit>=1.7.5           # Security linting
safety>=2.3.4           # Dependency security checks
flake8>=6.1.0           # Code linting
black>=23.12.0          # Code formatting
isort>=5.13.2           # Import sorting

# Performance & Load Testing (Python 3.12 compatible)
locust>=2.17.0          # Load testing